pyproject.toml
[project]
name = "chainreaction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.129.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.33",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
